refactor: restructured the model server packages
