+++ model-server/src/test/kotlin/org/modelix/model/server/JsonAPITest.kt
@@ -23,7 +23,12 @@ import io.ktor.server.websocket.*
 import org.json.JSONObject
 import org.modelix.authorization.installAuthentication
 import org.modelix.model.api.ITree
-import org.modelix.model.server.*
+import org.modelix.model.server.handlers.DeprecatedLightModelServer
+import org.modelix.model.server.handlers.asObjectList
+import org.modelix.model.server.handlers.buildJSONArray
+import org.modelix.model.server.handlers.buildJSONObject
+import org.modelix.model.server.store.InMemoryStoreClient
+import org.modelix.model.server.store.LocalModelClient
 import kotlin.test.*
 
 class JsonAPITest {
@@ -31,7 +36,7 @@ class JsonAPITest {
         application {
             installAuthentication(unitTestMode = true)
             install(WebSockets)
-            JsonModelServer(LocalModelClient(InMemoryStoreClient())).init(this)
+            DeprecatedLightModelServer(LocalModelClient(InMemoryStoreClient())).init(this)
         }
         block()
     }
